+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Only add to browser history when updating non-location settings
 - Add support for Leaflet styles
 - Document gazetteer configuration